Test file

+++ ft_calloc.c
@@ -17,7 +17,7 @@ void	*ft_calloc(size_t count, size_t size)
 	size_t	space;
 	void	*arr;
 
-	if (size != 0 && count > SIZE_MAX / size)
+	if (size != 0 && count > (UINT_MAX / size))
 		return (NULL);
 	space = count * size;
 	arr = malloc(space);
